Strong assumptions about package name.
With this modification, the compiler won't compile rules that are
defined on packages that contain upper case letters or import classes
that are themselves defined on packages with upper case letters. It's a
limitation that brings a lot of performance improvements, because the
isPackage method in the compiler becomes quite less expensive.